+++ src/App.js
@@ -11,6 +11,7 @@ import Checkout from './pages/checkout';
 import PageNotFound from './pages/pageNotFound';
 import PrivateRoute from './PrivateRoute';
 import Sales from './pages/sales';
+import Revenue from './pages/revenue';
 import AdminDashboard from './pages/adminDashboard';
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
@@ -29,6 +30,7 @@ function App() {
 				<PrivateRoute path="/checkout" exact component={Checkout} adminAuth={false}/>
 				<PrivateRoute path="/admin/dashboard" exact component={AdminDashboard} adminAuth={true}/>
 				<PrivateRoute path="/admin/sales" exact component={Sales} adminAuth={true}/>
+				<PrivateRoute path="/admin/revenue" exact component={Revenue} adminAuth={true}/>
 				<Route path="*" component={PageNotFound} />
 			</Switch>
 			<ToastContainer/>
