+++ src/components/UI/adminInventory/salesTable.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { toast } from 'react-toastify';
+import { API_URL } from '../../../constants/api';
 
 function SalesTable (props) {
     const [chartDatas, setChartDatas] = useState([]);
@@ -8,7 +9,7 @@ function SalesTable (props) {
     useEffect(() => {
         const fetchdata = async () => {
             try {
-                const response = await axios.get(`http://localhost:2001/admin/sales/${props.endpoint}`);
+                const response = await axios.get(`${API_URL}/admin/sales/${props.endpoint}`);
                 setChartDatas(response.data)
             } catch (error) {
                 toast.error(error.response.data.message || "Server Error", {
