+++ src/constants/api.js
@@ -1,6 +1,6 @@
 // Uncomment this part for Development
 const DOMAIN = "localhost"
-const PORT = "2000"
+const PORT = "2001"
 
 export const API_URL = `http://${DOMAIN}:${PORT}`;
