+++ src/pages/revenue.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Revenue = (props) => {
+    return (
+        <div className="text-center" style={{ width: "70vw" }}>
+            <div className="d-flex flex-row justify-content-between">
+                <div style={{ width: "40vw", height: "45vh", backgroundColor: "aquamarine" }}>
+                    fdssdfd
+                </div>
+                <div style={{ width: "25vw", height: "45vh", backgroundColor: "brown" }}>
+                    fsdfdfsfd
+                </div>
+            </div>
+        </div>
+    );
+}
+ 
+export default Revenue;
